backend/requirements.txt
# Dependencias principales
Django==5.0.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
django.db

# Base de datos
psycopg2-binary==2.9.9

# CORS y seguridad
django-cors-headers==4.3.1
django-filter==23.5

# Documentación API
drf-yasg==1.21.7

# Manipulación de imágenes
Pillow==10.1.0

# Tareas asíncronas
celery==5.3.6
redis==5.0.1

# Despliegue
whitenoise==6.6.0
gunicorn==21.2.0

# Desarrollo y testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
factory-boy==3.3.0
black==23.12.1
isort==5.13.2